portal: fix CollectorSummary Dapper type mismatch (FreshnessAgeSeconds long? → int?)
DATEDIFF(SECOND, ...) returns SQL int; Dapper could not materialize CollectorSummary
because the record had long? — 500 on GET /api/admin/collectors and /api/sim/collectors.

+++ portal/api/src/Mod.PortalApi/Features/Admin/Services/ICollectorService.cs
@@ -34,7 +34,7 @@ internal sealed record CollectorSummary(
     int? ReportedConfigVersion, int? ConfigVersion,
     DateTime? CertificateExpiresAt,
     int? BufferDepthEvents, int? BufferCapacityEvents,
-    long? FreshnessAgeSeconds);
+    int? FreshnessAgeSeconds);
 
 internal sealed record CollectorRow(
     Guid CollectorId, Guid TenantId, Guid SiteId, string Name,
